Assignment 7/Cloud_functions/clean_table/requirements.txt
google-cloud-bigquery
google-cloud-pubsub
pandas
db-dtypes